gh-475 Move TestProxyAggregationTask.main into own class
